Stability fixes regarding candle.end()

+++ Candle/src/usbDevice.cpp
@@ -115,7 +115,7 @@ bool UsbDevice::receive(int timeoutMs)
         std::cout << "Got " << std::dec << bytesReceived  << "bytes." <<std::endl;
         std::cout << rxBuffer << std::endl;
         for(int i = 0; i < bytesReceived; i++)
-            std::cout << std::hex << "0x" << (int)rxBuffer[i] << " ";
+            std::cout << std::hex << "0x" << (unsigned short)rxBuffer[i] << " ";
         std::cout << std::dec << std::endl << "#######################################################" << std::endl; 
     }
 #endif
